requirements/base.txt
# Base requirements for REJLERS Django Backend
# Core Django and DRF
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database - PostgreSQL & MongoDB
psycopg2-binary==2.9.7
dj-database-url==2.1.0
pymongo==4.6.0
djongo==1.3.6
motor==3.3.2

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
cryptography==41.0.7

# API Documentation
drf-spectacular==0.26.5
drf-spectacular[sidecar]==0.26.5

# Environment & Configuration
python-decouple==3.8
django-environ==0.11.2

# Validation & Serialization
marshmallow==3.20.1
django-filter==23.3

# File Handling & Storage
Pillow==10.0.1
django-storages==1.14.2
boto3==1.29.7

# Utilities
requests==2.31.0
python-slugify==8.0.1
django-extensions==3.2.3

# Time & Date
python-dateutil==2.8.2
pytz==2023.3

# Logging & Monitoring
sentry-sdk[django]==1.38.0

# Performance
django-redis==5.4.0
redis==5.0.1

# Content Management
django-ckeditor==6.7.0
django-taggit==4.0.0

# Email
django-anymail==10.2

# Internationalization
django-modeltranslation==0.18.11

# Development Tools (basic ones that might be needed in production)
factory-boy==3.3.0